Refactor: 슬라이스 컴포넌트 분리 - 상세페이지, 광고 Realated to: #6
